Honey collection is currently capped at 1/4 of the empty capacity of the stage.
Currently working to fix winter. For now, I'm using the interp function to get predicted honey, but I've capped the stored honey at 1/4 of the vacant cells available that day.  This allows us to see what is going on reasonably well- honey IS being consumed, but there is still this weird jump in collection. For now, I'm letting it go, and working on not having them all die out in winter.